% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{directed_mst_visualization}
\alias{directed_mst_visualization}
\title{Construct km-curve}
\usage{
directed_mst_visualization(
  directed_mst,
  TF,
  pdf_width = 8,
  pdf_height = 7,
  png_width = 2400,
  png_height = 2100,
  png_resolution = 300,
  line_transparency = 1,
  line_thickness = 1,
  parent_node_size = 18,
  intermediate_node_size = 10,
  lead_node_size = 5,
  text_size = 0.65,
  text_type = 2,
  text_centrality = -0.5,
  text_diagnol = -pi/2,
  node_color = "white",
  node_frame_color = "grey",
  arrow_size = 0.6,
  aspect_ratio = 0.6
)
}
\arguments{
\item{directed_mst}{Genes selected for km construction}

\item{TF}{RDS with survival & patient info}

\item{pdf_width}{As parameter name}

\item{pdf_height}{As parameter name}

\item{png_width}{As parameter name}

\item{png_height}{As parameter name}

\item{png_resolution}{As parameter name}

\item{line_transparency}{As parameter name}

\item{line_thickness}{As parameter name}

\item{parent_node_size}{As parameter name}

\item{intermediate_node_size}{As parameter name}

\item{lead_node_size}{As parameter name}

\item{text_size}{As parameter name}

\item{text_type}{As parameter name}

\item{text_centrality}{As parameter name}

\item{text_diagnol}{As parameter name}

\item{node_color}{As parameter name}

\item{node_frame_color}{As parameter name}

\item{arrow_size}{As parameter name}

\item{aspect_ratio}{As parameter name}
}
\value{
GRN graph
}
\description{
Construct km-curve
}
