% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{dynamic_weight_initialization}
\alias{dynamic_weight_initialization}
\title{Caculate weight based on pseudotime}
\usage{
dynamic_weight_initialization(meta_data, imputed_matrix)
}
\arguments{
\item{meta_data}{Data frame that has all calculated value we calculate before}

\item{imputed_matrix}{Imputed gene matrix}
}
\value{
Peak information & correlation informaiton
}
\description{
Caculate weight based on pseudotime
}
