% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{find_best_k_kmean}
\alias{find_best_k_kmean}
\title{Use Silhouette find best k for k mean clustering}
\usage{
find_best_k_kmean(df_umap, test_k = 10, nstart = 10)
}
\arguments{
\item{df_umap}{UMAP data frame}

\item{test_k}{Number of k you wanna test}

\item{nstart}{K-mean clustering parameter}
}
\value{
Best k
}
\description{
Use Silhouette find best k for k mean clustering
}
