% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{find_peak}
\alias{find_peak}
\title{Find peak of loess}
\usage{
find_peak(loess_fit, x, y)
}
\arguments{
\item{loess_fit}{The loess construct before}

\item{x}{Gene expression}

\item{y}{Seurat Score}
}
\value{
Peak of loess
}
\description{
Find peak of loess
}
