% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{generate_penalty_score}
\alias{generate_penalty_score}
\title{Caculate penalty score for each tf}
\usage{
generate_penalty_score(
  genie_matrix,
  df_tpm,
  quantile_por = 0.9,
  threshold = 0.5,
  alpha = 2,
  amplify = 3
)
}
\arguments{
\item{genie_matrix}{The GENIE3 weight matrix}

\item{df_tpm}{Expression matrix (original)}

\item{quantile_por}{Threshold set for highly expression cluster group}

\item{threshold}{Threshold set for portion}

\item{alpha}{Penalty degree}

\item{amplify}{Amplification degree of portion}
}
\value{
Penalty score of batch effect
}
\description{
Caculate penalty score for each tf
}
