% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{hvg_selection}
\alias{hvg_selection}
\title{Use loess select high variance gene}
\usage{
hvg_selection(df, p_value = 0.2, span = 0.3)
}
\arguments{
\item{df}{Data frame with gene expression}

\item{p_value}{Threshold for variance gene}

\item{span}{Loess}
}
\value{
All high variance gene
}
\description{
Use loess select high variance gene
}
