% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM_analysis.R
\name{km_curve}
\alias{km_curve}
\title{Construct km-curve}
\usage{
km_curve(signature_genes, patient_RDS, TF, cut = "median")
}
\arguments{
\item{signature_genes}{Genes selected for km construction}

\item{patient_RDS}{RDS with survival & patient info}

\item{TF}{Selected TF for km construction}

\item{cut}{Default median, otherwise optimal}
}
\value{
km-curve
}
\description{
Construct km-curve
}
