% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{km_curve_visualization}
\alias{km_curve_visualization}
\title{Construct km-curve}
\usage{
km_curve_visualization(
  km_fit,
  df,
  Title,
  pdf_width = 10,
  pdf_height = 6.5,
  png_width = 2400,
  png_height = 1600,
  png_resolution = 300
)
}
\arguments{
\item{km_fit}{The km curve constructed earlier}

\item{df}{Survival df}

\item{Title}{Usually TF}

\item{pdf_width}{As parameter name}

\item{pdf_height}{As parameter name}

\item{png_width}{As parameter name}

\item{png_height}{As parameter name}

\item{png_resolution}{As parameter name}
}
\value{
visualized km-curve
}
\description{
Construct km-curve
}
