% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layered_graph.R
\name{layered_graph}
\alias{layered_graph}
\title{Construct GRN for selected tf}
\usage{
layered_graph(TF, weight_matrix, layer1_top = 10, layer2_top = 5)
}
\arguments{
\item{TF}{Selected TF for GRN construction}

\item{weight_matrix}{Weight Matrix's name}

\item{layer1_top}{How many TF you want to see in first layer}

\item{layer2_top}{How many TF you want to see in second layer}
}
\value{
Layered graph
}
\description{
Construct GRN for selected tf
}
