% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{load_pseudotime}
\alias{load_pseudotime}
\title{Calculate pseudotime for cell progression}
\usage{
load_pseudotime(RDS_file, nstart = 10)
}
\arguments{
\item{RDS_file}{RDS file with tpm}

\item{nstart}{K-mean clustering parameter}
}
\value{
Pseudotime
}
\description{
Calculate pseudotime for cell progression
}
