% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{load_state_score}
\alias{load_state_score}
\title{Calculate state score as pseudo-pseudotime for cell progression}
\usage{
load_state_score(RDS_file, gene_list)
}
\arguments{
\item{RDS_file}{RDS file with tpm}

\item{gene_list}{A list of gene that represent your selection of cell type}
}
\value{
Pseudo-pseudotime
}
\description{
Calculate state score as pseudo-pseudotime for cell progression
}
