% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{pca_knn_imputation}
\alias{pca_knn_imputation}
\title{Use kNN impute missing value}
\usage{
pca_knn_imputation(df, hvg_list, k = 5)
}
\arguments{
\item{df}{Data frame with gene expression}

\item{hvg_list}{High variance gene calculate before}

\item{k}{The number of k you choose for knn imputation}
}
\value{
Imputated expression data frame
}
\description{
Use kNN impute missing value
}
