% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{pca_plot_visualization}
\alias{pca_plot_visualization}
\title{Construct PCA plot}
\usage{
pca_plot_visualization(
  target_gene,
  RDS_file,
  PCA = 10,
  cut = 0.9,
  background_dot_transparency = 0.5,
  top10_dot_transparency = 0.8,
  background_dot_size = 1,
  top10_dot_size = 1.5,
  eclipse_coverage = 0.95,
  eclipse_line_width = 1.2,
  axis_title_size = 12,
  plot_title_size = 14,
  pdf_width = 10,
  pdf_height = 6.5,
  png_width = 2400,
  png_height = 1600,
  png_resolution = 300
)
}
\arguments{
\item{target_gene}{Gene name you want to see distribution}

\item{RDS_file}{RDS file with tpm}

\item{PCA}{Number of PCA, 10 is sufficient}

\item{cut}{Quantile you set}

\item{background_dot_transparency}{As parameter name}

\item{top10_dot_transparency}{As parameter name}

\item{background_dot_size}{As parameter name}

\item{top10_dot_size}{As parameter name}

\item{eclipse_coverage}{As parameter name}

\item{eclipse_line_width}{As parameter name}

\item{axis_title_size}{As parameter name}

\item{plot_title_size}{As parameter name}

\item{pdf_width}{As parameter name}

\item{pdf_height}{As parameter name}

\item{png_width}{As parameter name}

\item{png_height}{As parameter name}

\item{png_resolution}{As parameter name}
}
\value{
visualized pca plot
}
\description{
Construct PCA plot
}
