% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{penalty_score_batch_effect}
\alias{penalty_score_batch_effect}
\title{Caculate penalty score based on expression's dispersion}
\usage{
penalty_score_batch_effect(
  tf_expression,
  pca_coordinates,
  global_disperse,
  quantile_por = 0.9,
  threshold = 0.5,
  alpha = 2,
  amplify = 3
)
}
\arguments{
\item{tf_expression}{Data frame that has all calculated value we calculate before}

\item{pca_coordinates}{Imputed gene matrix}

\item{global_disperse}{Global dispersion}

\item{quantile_por}{Threshold set for highly expression cluster group}

\item{threshold}{Threshold set for portion}

\item{alpha}{Penalty degree}

\item{amplify}{Amplification degree of portion}
}
\value{
Penalty score of batch effect
}
\description{
Caculate penalty score based on expression's dispersion
}
