% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM_analysis.R
\name{select_co_factor}
\alias{select_co_factor}
\title{Select co-factor for selected tf}
\usage{
select_co_factor(weight_matrix, TF, top_percentage = 0.9)
}
\arguments{
\item{weight_matrix}{Weight Matrix's name}

\item{TF}{Selected TF for km construction}

\item{top_percentage}{Treshold for co-factor}
}
\value{
co-factor
}
\description{
Select co-factor for selected tf
}
