% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{top_TRweight_genes}
\alias{top_TRweight_genes}
\title{Get top expressed TF based on previously calculated weight}
\usage{
top_TRweight_genes(weight_matrix)
}
\arguments{
\item{weight_matrix}{Weight matrix you calculated before}
}
\value{
Top expressed TF
}
\description{
Get top expressed TF based on previously calculated weight
}
