% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{violin_plot_visualization}
\alias{violin_plot_visualization}
\title{Visualize gene expression across patient samples using violin plots}
\usage{
violin_plot_visualization(
  target_genes,
  RDS_file,
  Title,
  Subtitle,
  output_name,
  pdf_width = 8,
  pdf_height = 5,
  png_width = 2400,
  png_height = 1500,
  png_resolution = 300
)
}
\arguments{
\item{target_genes}{Character vector of gene names to visualize.}

\item{RDS_file}{Path to an RDS file containing TPM expression data and sample information.}

\item{Title}{Title of the plot.}

\item{Subtitle}{Subtitle of the plot.}

\item{output_name}{File name prefix used for exporting the figure.}

\item{pdf_width}{Width of the exported PDF.}

\item{pdf_height}{Height of the exported PDF.}

\item{png_width}{Width of the exported PNG.}

\item{png_height}{Height of the exported PNG.}

\item{png_resolution}{Resolution of the exported PNG.}
}
\value{
A violin plot showing gene expression across patient samples.
}
\description{
Visualize gene expression across patient samples using violin plots
}
