% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{weight_construction}
\alias{weight_construction}
\title{Caculate weight by calling previous function one by one}
\usage{
weight_construction(genie_matrix, dynamic_matrix, alpha)
}
\arguments{
\item{genie_matrix}{Genie weight matrix that generated from GENIE3}

\item{dynamic_matrix}{Dynamic weight matrix that contain all necessary information for dynamic weight calculation}

\item{alpha}{Time penalty degree}
}
\value{
Calculated weight matrix
}
\description{
Caculate weight by calling previous function one by one
}
