% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{TR_weight}
\alias{TR_weight}
\title{Caculate weight by calling previous function one by one}
\usage{
TR_weight(
  RDS_file,
  gene_list,
  genie_file = NULL,
  batch_effect = TRUE,
  alpha = 0.5,
  nTree = 500,
  nCores = 1
)
}
\arguments{
\item{RDS_file}{RDS file with treated T/F, tpm}

\item{gene_list}{List of gene selected for scoring}

\item{genie_file}{If genie calculated before, import file}

\item{batch_effect}{Whether select slingshot or seurat}

\item{alpha}{Penalty degree}

\item{nTree}{Number of tree for genie3}

\item{nCores}{Number of core for genie3}
}
\value{
Weight matrix and other additional info might be useful
}
\description{
Caculate weight by calling previous function one by one
}
