% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{violin_plot_visualization}
\alias{violin_plot_visualization}
\title{Construct PCA plot}
\usage{
violin_plot_visualization(
  target_genes,
  RDS_file,
  Title,
  Subtitle,
  output_name,
  pdf_width = 8,
  pdf_height = 5,
  png_width = 2400,
  png_height = 1500,
  png_resolution = 300
)
}
\arguments{
\item{RDS_file}{RDS file with tpm}

\item{Title}{The name you want to set for title}

\item{Subtitle}{The name you want to set for subtitle}

\item{output_name}{The name you want to save this file as}

\item{pdf_width}{As parameter name}

\item{pdf_height}{As parameter name}

\item{png_width}{As parameter name}

\item{png_height}{As parameter name}

\item{png_resolution}{As parameter name}

\item{target_gene}{Gene name you want to see distribution}
}
\value{
visualized pca plot
}
\description{
Construct PCA plot
}
